+++ src/App.tsx
@@ -1,3 +1,4 @@
+import CartPage from './pages/CartPage';
 import MainPage from './pages/MainPage';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 
@@ -6,6 +7,7 @@ const App = () => {
     <BrowserRouter basename={process.env.PUBLIC_URL}>
       <Routes>
         <Route path="/" element={<MainPage />} />
+        <Route path="/cart" element={<CartPage />} />
       </Routes>
     </BrowserRouter>
   );
