+++ src/components/main/ProductList.tsx
@@ -1,10 +1,10 @@
 import { useRecoilState } from 'recoil';
 import { styled } from 'styled-components';
 import ProductItem from './ProductItem';
-import { useFetchData } from '../hooks/useFetchData';
-import { Product } from '../types';
-import { MOCK_DATA_URL } from '../constants/url';
-import { productListState } from '../store/ProductListState';
+import { useFetchData } from '../../hooks/useFetchData';
+import { Product } from '../../types';
+import { MOCK_DATA_URL } from '../../constants/url';
+import { productListState } from '../../store/ProductListState';
 
 const ProductList = () => {
   const [productList, setProductList] = useRecoilState<Product[]>(productListState);
